Refactor to non-mutating function structure.
